Initial idea spike
The idea came from @tehviking

In order to use this we have to add it to your .irbrc. Or to play around
you can open irb and then

```ruby
  require "irb_affirmations"
  irb
```

I wanted to get this out there before I fixed the issue of needing to
reload irb.

Amos King @adkron <amos@binarynoggin.com>